Keep a linked list sorted
The four linked lists of appointment panel items (appointments, recurring
appointments, events, recurring events) are kept sorted by inserting elements
in order, either when they are first loaded from disk or when new are added.
The ordering is by start time (numerical) and description (alphabetical).

The user is allowed to change start time (for appointments only) as well as
description (for events also). A change is committed directly to the item
(unlike for cut/paste where an item is deleted and then inserted).
This may break the order. (Apparently the order property is only used when
events are loaded from the evenlist into the day_item vector (see
LLIST_FIND_FOREACH_CONT).)

A general linked list function is introduced that will reorder a list after a
list element has changed.