Eliminate double list search
When searching for the predecessor of a list element, the list element itself
was searched for before the predecessor, in order to get the list item used
in the predecessor search (in llist_reorder: llist_find_first() followed by
llist_prev()).

The double search has been eliminated by extending the predecessor search to
use the list element (*data) when searching; when the predecessor has
been found, the element itself is then the successor.