Ensure the appointment list is always sorted
When an appointment is moved, it may no longer show up in the notification bar.
The reason is that the search for the next appointment assumes a sorted list
of appointments. After load/reload of data all lists are ordered. After a move
(or change of start time or description) the list may be unsorted because the
changes are performed directly in the in-memory list.

A new function apoint_reorder(), based on llist_reorder(), sorts the in-memory
list and is called after the changes mentioned.

Unrelated cleanup in apoint_check_next(): removed function return value;
changed long to time_t.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>