Reimplementation of notification main thread
The update() routine has been incorporated into the main loop and the global
variable notify (and its type declaration) has been eliminated.

Because the main loop is run once every second, processing has been reduced as
much as possible. Update of the next appointment is only done once a minute (on
the minute). For that purpose the notify_app structure has been extended with an
element, update, which saves the time (before start) when the next update should
take place.

The nbar mutex is not used by the main thread and has been removed from the
cleanup handler. The notify_app mutex has been added.

Calculations have either been moved outside the main loop or are only repeated
when necessary.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>