Daemon review and revision
The daemon now immediately does a fork() and return in the parent. In this way
the main calcurse process continues after starting the daemon instead of exiting
through the daemon.

In exit_calcurse(), the daemon is now started before rather than after freeing
user data. Hence the daemon does not have to load anything, but instead can free
unused data. Also, the notify_app data are then available so that the daemon
will not launch an already notified appointment.

With the --daemon option (command line start) the daemon is not started if
calcurse is running, and only appointments are loaded.

The daemon can handle the next appointment itself and does not have to launch
the next app thread (notify_check_next_app()). Function notify_get_next_bkgd()
has been removed from notify.c and integrated in the main loop. Because the
daemon is running alone, there is no need for the notify_app mutex.

The readability of log messages has been improved, and notify_needs_reminder()
has been eliminated.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>