Disable user-defined shell command
Notification, warning, reminder: different words for the existing behaviour.
If nbar.cntdwn (notification.warning) is set to zero in the notify menu, the
visual alert (blinking text in the notification bar) as well as the audible
(default user-defined shell command) is turned off.

The shell command may be independently disabled by inserting a '#' at the start
of the command string. Note that no child process is spawned.

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>