Format specifiers for notification shell command
In the notification command of the notify configuration menu, format specifiers
are implemented which relate to the next appointment and are expanded before the
string is handed to the shell.

The specifiers are %m, %d and %t, which are expanded, respectively, into the
message, the date and the time of the next appointment. The date format is that
used for the notification bar; the time format is fixed as "%H:%M".

Signed-off-by: Lars Henriksen <LarsHenriksen@get2net.dk>